Fix build to compile on windows
