+++ pkg/config/config_nix.go
@@ -14,3 +14,17 @@ const (
 	defaultJMXPipePath          = "/opt/datadog-agent/run"
 	defaultSyslogURI            = "unixgram:///dev/log"
 )
+
+// GetSyslogURI returns the configured/default syslog uri
+func GetSyslogURI() string {
+	enabled := Datadog.GetBool("log_to_syslog")
+	uri := Datadog.GetString("syslog_uri")
+
+	if enabled {
+		if uri == "" {
+			uri = defaultSyslogURI
+		}
+	}
+
+	return uri
+}
