+++ pkg/config/config_windows.go
@@ -5,6 +5,10 @@
 
 package config
 
+import (
+	log "github.com/cihub/seelog"
+)
+
 const (
 	defaultConfdPath            = "c:\\programdata\\datadog\\conf.d"
 	defaultAdditionalChecksPath = "c:\\programdata\\datadog\\checks.d"
@@ -13,5 +17,10 @@ const (
 )
 
 func GetSyslogURI() string {
+	enabled := Datadog.GetBool("log_to_syslog")
+
+	if enabled {
+		log.Infof("logging to syslog is not available on windows.")
+	}
 	return ""
 }
